Move packages to /packages directory
- Use the same tsconfig
- Add a GitHub Actions job to install and build the packages
