+++ .github/workflows/build-packages.yml
@@ -1,0 +1,29 @@
+name: Build packages
+
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v1
+
+      - name: Setup Node.js 12.x
+        uses: actions/setup-node@v1
+        with:
+          node-version: 12.x
+
+      - name: Install and build @excalidraw/excalidraw
+        run: |
+          npm ci --prefix packages/embed
+          npm run pack --prefix packages/embed
+
+      - name: Install and build @excalidraw/utils
+        run: |
+          npm ci --prefix packages/embed
+          npm run pack --prefix packages/embed
