+++ .gitignore
@@ -15,3 +15,4 @@ yarn.lock
 .idea
 dist/
 .eslintcache
+*.tgz
