+++ packages/embed/.gitignore
