+++ packages/embed/package.json
