--- src/excalidraw-utils/index.js
@@ -1,1 +1,0 @@
-export { exportToBlob, exportToSvg } from "../utils";
