+++ src/utils.ts
@@ -1,8 +1,3 @@
-import {
-  exportToCanvas as _exportToCanvas,
-  exportToSvg as _exportToSvg,
-} from "./scene/export";
-import { getDefaultAppState } from "./appState";
 import { AppState } from "./types";
 import { getZoomOrigin } from "./scene";
 import {
@@ -10,8 +5,7 @@ import {
   FONT_FAMILY,
   WINDOWS_EMOJI_FALLBACK_FONT,
 } from "./constants";
-import { ExcalidrawElement, FontFamily, FontString } from "./element/types";
-import { getNonDeletedElements } from "./element";
+import { FontFamily, FontString } from "./element/types";
 
 export const SVG_NS = "http://www.w3.org/2000/svg";
 
@@ -295,87 +289,4 @@ export const findLastIndex = <T>(
     }
   }
   return -1;
-};
-
-// Export functions
-
-type ExportOpts = {
-  elements: readonly ExcalidrawElement[];
-  appState?: Omit<AppState, "offsetTop" | "offsetLeft">;
-  getDimensions: (
-    width: number,
-    height: number,
-  ) => { width: number; height: number; scale: number };
-};
-
-const exportToCanvas = ({
-  elements,
-  appState = getDefaultAppState(),
-  getDimensions = (width, height) => ({ width, height, scale: 1 }),
-}: ExportOpts) => {
-  return _exportToCanvas(
-    getNonDeletedElements(elements),
-    { ...appState, offsetTop: 0, offsetLeft: 0 },
-    {
-      exportBackground: appState.exportBackground ?? true,
-      viewBackgroundColor: appState.viewBackgroundColor ?? "#FFF",
-      shouldAddWatermark: appState.shouldAddWatermark ?? false,
-    },
-    (width, height) => {
-      const canvas = document.createElement("canvas");
-      const ret = getDimensions(width, height);
-
-      canvas.width = ret.width;
-      canvas.height = ret.height;
-
-      return canvas;
-    },
-  );
-};
-
-export const exportToBlob = (
-  opts: ExportOpts & {
-    mimeType?: string;
-    quality?: number;
-  },
-): Promise<Blob | null> => {
-  const canvas = exportToCanvas(opts);
-
-  let { mimeType = "image/png", quality } = opts;
-
-  if (mimeType === "image/png" && typeof quality === "number") {
-    console.warn(`"quality" will be ignored for "image/png" mimeType`);
-  }
-
-  if (mimeType === "image/jpg") {
-    mimeType = "image/jpeg";
-  }
-
-  quality = quality ? quality : /image\/jpe?g/.test(mimeType) ? 0.92 : 0.8;
-
-  return new Promise((resolve) => {
-    canvas.toBlob(
-      (blob) => {
-        resolve(blob);
-      },
-      mimeType,
-      quality,
-    );
-  });
-};
-
-export const exportToSvg = ({
-  elements,
-  appState = getDefaultAppState(),
-  exportPadding,
-  metadata,
-}: ExportOpts & {
-  exportPadding?: number;
-  metadata?: string;
-}): SVGSVGElement => {
-  return _exportToSvg(getNonDeletedElements(elements), {
-    ...appState,
-    exportPadding,
-    metadata,
-  });
 };
